Back merge of the trunk - 0-persistence files add
git-svn-id: svn+ssh://scm.gforge.inria.fr/svnroot/gudhi@550 636b058d-ea47-450e-bf9e-a15bfbe3eedb

Former-commit-id: d2fe05be71ffa4e8841b41996959a0f3d9bd136e

+++ branches/0_Persistence/src/0_persistence/example/density_then_0_pers.cpp
@@ -96,7 +96,7 @@ int main(int argc, char *argv[]){
 	cout << "Dimension: " << dim << endl;
       }
       // else check that dimension is preserved
-      else if (dim != row.size()) {
+      else if (dim != static_cast<int>(row.size())) {
 	cerr << "Error: mismatched dimension in " 
 	     << input_file_name << " at line " << (nb_points+1) << endl;
 	return -1;
